requirements.txt
absl-py==0.7.1
amqp==2.4.2
astor==0.7.1
backports.functools-lru-cache==1.6.1
backports.weakref==1.0.post1
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
contextlib2==0.5.5
cycler==0.10.0
Cython==0.29.7
Django==1.11
dnspython==1.16.0
enum34==1.1.10
eureka-client==0.6.2
funcsigs==1.0.2
futures==3.3.0
gast==0.2.2
grpcio==1.20.1
h5py==2.9.0
idna==2.8
imageai==2.1.5
imutils==0.5.2
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
kombu==4.5.0
lxml==4.3.3
Markdown==3.1
matplotlib==2.2.5
mock==3.0.4
numpy==1.16.3
opencv-python==4.1.0.25
pandas==0.24.2
Pillow==6.0.0
protobuf==3.7.1
psycopg2-binary==2.8.2
py-eureka-client==0.5.7
pyparsing==2.4.0
pytesseract==0.2.6
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
requests==2.21.0
scikit-learn==0.20.4
scipy==1.2.1
six==1.12.0
sklearn==0.0
subprocess32==3.5.4
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
tqdm==4.32.1
urllib3==1.24.3
vine==1.3.0
Werkzeug==0.15.2
